requirements.txt
weatherhat==0.0.1
python-dotenv
paho-mqtt